+++ src/SW_Flow_lib.c
@@ -474,19 +474,18 @@ used for transpiration calculations.
     the simulated site's input values
 @param[in] SW_SiteSim Struct of type SW_SITE_SIM describing the simulated site's
     simulation values
-@param[in] n_tr_rgns Array of n_lyrs elements of transpiration regions that each
-    soil layer belongs to.
-@param[in] n_layers Number of soil layers.
-@param[in] tr_regions Number of layer regions used in weighted average,
-    typically 3;<BR> to represent shallow, mid, & deep depths to compute
-    transpiration rate.
+@param[in] n_tr_rgns Number of transpiration regions.
+@param[in] n_layers Number of soil layers with roots
+    (for a specific plant functional type)
+@param[in] tr_regions Array with the transpiration region assigned to each
+    soil layer (for a specific plant functional type)
 @param[in] swc Soilwater content in each layer before drainage
     (m<SUP>3</SUP>H<SUB>2</SUB>O).
 @param[in] VegType Current vegetation
 @param[out] LogInfo Holds information on warnings and errors
 
 @sideeffect *swp_avg Weighted average of soilwater potential and transpiration
-coefficients (-bar).
+coefficients (-bar). NAN if all transpiration coefficients are 0.
 */
 void transp_weighted_avg(
     double *swp_avg,
@@ -527,7 +526,8 @@ void transp_weighted_avg(
     double swp;
     double sumco;
 
-    *swp_avg = 0;
+    *swp_avg = NAN;
+
     for (r = 1; r <= n_tr_rgns; r++) {
         swp = sumco = 0.0;
 
@@ -546,10 +546,15 @@ void transp_weighted_avg(
             }
         }
 
-        swp /= GT(sumco, 0.) ? sumco : 1.;
-
-        /* use smallest weighted average of regions */
-        (*swp_avg) = (r == 1) ? swp : fmin(swp, (*swp_avg));
+        /* Ignore swp of current regions if sumco is 0
+           (i.e., all transp_coeff in current region are zero)
+           --> swp_avg could potentially remain NAN
+        */
+        if (GT(sumco, 0.)) {
+            swp /= sumco;
+            /* use smallest (i.e., wettest) weighted average across regions */
+            (*swp_avg) = (r == 1) ? swp : fmin(swp, (*swp_avg));
+        }
     }
 }
 
@@ -771,7 +776,7 @@ void pot_soil_evap_bs(
 
 Based on equations from Parton 1978. @cite Parton1978
 
-@param *bstrate This is the bare soil evaporation loss rate (cm/day).
+@param *bstrate This is the potential transpiration loss rate (cm/day).
 @param swpavg Weighted average of soil water potential (-bar).
 @param biolive Living biomass (%).
 @param biodead Dead biomass (%).
@@ -794,7 +799,8 @@ Based on equations from Parton 1978. @cite Parton1978
 @param co2_wue_multiplier Water-usage efficiency multiplier (CO<SUB>2</SUB>
 ppm).
 
-@sideeffect *bstrate Updated bare soil evaporation loss rate (cm/day).
+@sideeffect *bstrate Updated potential transpiration loss rate (cm/day)
+    which is 0 if \p biolive is 0 or if \p swpavg is missing.
 */
 void pot_transp(
     double *bstrate,
@@ -844,7 +850,7 @@ void pot_transp(
     double par2;
     double shadeaf;
 
-    if (LE(biolive, 0.)) {
+    if (LE(biolive, 0.) || missing(swpavg)) {
         *bstrate = 0.;
 
     } else {